backend/requirements.txt
# Flask Web Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1

# MySQL Database
mysql-connector-python==8.2.0

# LangChain & AI
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.5
google-generativeai==0.8.3

# Embeddings & Vector Store
langchain-huggingface==0.1.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0
huggingface-hub==0.26.2

# Utilities
pydantic==2.10.3